Cover HIPAA §(J) account numbers; split hipaa.rs by posture
Before this change HIPAA Safe Harbor / Limited Data Set / Expert
Determination scoped `bank_account` alone under §(J). Real §(J) —
"account numbers" per §164.514(b)(2)(i)(J) and §164.514(e)(2)(x) —
covers every individual-linked account identifier the pattern set
detects: also `iban`, `payment_card`, and (under the §(R)
catch-all reading) `crypto_address`.

Add a `HipaaAccountNumbers` enum on `PolicyTemplate::HipaaDeidentification`
picking the account label scope:

- `Standard` (default): `bank_account` + `iban` + `payment_card`.
  The traditional §(J) reading; matches every de-ID vendor's
  Safe Harbor preset.
- `Extended`: adds `crypto_address`. Deployments where wallet
  addresses can appear in patient-facing artifacts opt in.

Both tiers apply uniformly across Safe Harbor, LDS, and Expert
Determination — §(J) scope is a property of the identifier list,
not the removal method. Threaded through `hipaa::template()` and
the three per-method builders. Caller-side change is one new
`accounts` field on the wire (`#[serde(default)]` so callers
already sending `{"method": "safe_harbor"}` continue to work).

While in there, split the 800-line `hipaa.rs` into a `hipaa/`
folder — one file per posture (`safe_harbor.rs`, `limited_data_set.rs`,
`expert_determination.rs`), one file for the account-tier enum
(`account_numbers.rs`), shared items and tests in `mod.rs`. Each
posture owns its own constants and UUIDs; Expert Determination
reaches into `safe_harbor` for the shared 18-identifier label set.

New tests:

- `standard_accounts_cover_the_traditional_section_j_set` — pins
  the Standard tier.
- `extended_accounts_add_crypto_to_the_standard_set` — pins
  Extended.
- `account_tier_lands_in_every_hipaa_method` — verifies the tier
  widens every method's policy scope.
