+++ crates/nvisy-template/src/template/mod.rs
@@ -16,7 +16,7 @@ use semver::Version;
 use serde::{Deserialize, Serialize};
 
 pub use self::gdpr::GdprArticle9Treatment;
-pub use self::hipaa::HipaaDeidMethod;
+pub use self::hipaa::{HipaaAccountNumbers, HipaaDeidMethod};
 pub use self::pci::{PciDssPart, PciPanRender};
 
 /// A regulatory posture packaged as engine-ready data.
